CI: Restructure imports for tests, add path for cache in git workflow
- Update GitHub Actions workflows for E2E and unit tests
- Modify backend package.json and test files
