+++ .github/workflows/run-e2e-tests.yaml
@@ -15,6 +15,7 @@ jobs:
           with:
             node-version: 20
             cache: "npm"
+            cache-dependency-path: ./backend/package-lock.json
      
         - run: npm ci
