+++ backend/package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "scripts": {
     "dev": "nodemon src/server.js",
-    "test:e2e": "node --test src/tests/controllers",
-    "test:unit": "node --test src/tests/units"
+    "test:e2e": "node --test tests/controllers",
+    "test:unit": "node --test tests/units"
   },
   "devDependencies": {
     "nodemon": "^3.1.7",
